@@ -9,7 +9,6 @@ import (
 	"fmt"
 	"html/template"
 	"io/fs"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strconv"
@@ -20,7 +19,6 @@ import (
 	vcutils "github.com/trimble-oss/tierceron/trcconfigbase/utils"
 	"github.com/trimble-oss/tierceron/trcx/xutil"
 	"github.com/trimble-oss/tierceron/validator"
-	"github.com/trimble-oss/tierceron/vaulthelper/kv"
 	helperkv "github.com/trimble-oss/tierceron/vaulthelper/kv"
 
 	eUtils "github.com/trimble-oss/tierceron/utils"
@@ -42,7 +40,7 @@ type writeCollection struct {
 
 var templateWritten map[string]bool
 
-func GetTemplateParam(config *eUtils.DriverConfig, mod *kv.Modifier, filePath string, paramWanted string) (string, error) {
+func GetTemplateParam(config *eUtils.DriverConfig, mod *helperkv.Modifier, filePath string, paramWanted string) (string, error) {
 
 	templateEncoded, err := vcutils.GetTemplate(config, mod, filePath)
 	if err != nil {
@@ -87,7 +85,7 @@ func SeedVault(config *eUtils.DriverConfig) error {
 	config.Log.SetPrefix("[SEED]")
 	config.Log.Printf("Seeding vault from seeds in: %s\n", config.StartDir[0])
 
-	files, err := ioutil.ReadDir(config.StartDir[0])
+	files, err := os.ReadDir(config.StartDir[0])
 
 	if len(config.FileFilter) == 1 && config.FileFilter[0] == "nest" {
 		dynamicPathFilter := ""
@@ -163,7 +161,7 @@ func SeedVault(config *eUtils.DriverConfig) error {
 			templateParam = strings.Replace(templateParam, "ENV", config.EnvRaw, -1)
 			wd, err := os.Getwd()
 			if err != nil {
-				eUtils.LogErrorObject(config, errors.New("Could not get working directory for cert existence verification."), false)
+				eUtils.LogErrorObject(config, errors.New("could not get working directory for cert existence verification"), false)
 				continue
 			}
 
@@ -180,7 +178,7 @@ func SeedVault(config *eUtils.DriverConfig) error {
 		if len(tempPaths) > 0 {
 			templatePaths = tempPaths
 		} else {
-			return eUtils.LogErrorAndSafeExit(config, errors.New("No valid cert files were located."), -1)
+			return eUtils.LogErrorAndSafeExit(config, errors.New("no valid cert files were located"), -1)
 		}
 		_, _, seedData, errGenerateSeeds := xutil.GenerateSeedsFromVaultRaw(config, true, templatePaths)
 		if errGenerateSeeds != nil {
@@ -217,7 +215,7 @@ func SeedVault(config *eUtils.DriverConfig) error {
 			config.Log.Println("\tStepping into: " + envDir.Name())
 
 			if config.DynamicPathFilter != "" {
-				sectionConfigFiles, err := ioutil.ReadDir(config.StartDir[0] + "/" + envDir.Name() + "/" + config.DynamicPathFilter)
+				sectionConfigFiles, err := os.ReadDir(config.StartDir[0] + "/" + envDir.Name() + "/" + config.DynamicPathFilter)
 				if err != nil {
 					config.Log.Printf("Seed Sections Couldn't read into: %s \n", config.DynamicPathFilter)
 				}
@@ -239,11 +237,11 @@ func SeedVault(config *eUtils.DriverConfig) error {
 				continue
 			}
 
-			var filesSteppedInto []fs.FileInfo
+			var filesSteppedInto []fs.DirEntry
 			if indexedEnvNot {
-				filesSteppedInto, err = ioutil.ReadDir(config.StartDir[0] + "/" + envDir.Name() + "/" + suffix)
+				filesSteppedInto, err = os.ReadDir(config.StartDir[0] + "/" + envDir.Name() + "/" + suffix)
 			} else {
-				filesSteppedInto, err = ioutil.ReadDir(config.StartDir[0] + "/" + envDir.Name())
+				filesSteppedInto, err = os.ReadDir(config.StartDir[0] + "/" + envDir.Name())
 			}
 			eUtils.LogErrorObject(config, err, true)
 
@@ -279,7 +277,7 @@ func SeedVault(config *eUtils.DriverConfig) error {
 					if eUtils.OnlyBasePtr {
 						continue
 					}
-					projectDirectories, err := ioutil.ReadDir(config.StartDir[0] + "/" + envDir.Name() + "/" + fileSteppedInto.Name())
+					projectDirectories, err := os.ReadDir(config.StartDir[0] + "/" + envDir.Name() + "/" + fileSteppedInto.Name())
 					if err != nil {
 						config.Log.Printf("Projects Couldn't read into: %s \n", fileSteppedInto.Name())
 					}
@@ -297,7 +295,7 @@ func SeedVault(config *eUtils.DriverConfig) error {
 								continue
 							}
 						}
-						sectionNames, err := ioutil.ReadDir(config.StartDir[0] + "/" + envDir.Name() + "/" + fileSteppedInto.Name() + "/" + projectDirectory.Name())
+						sectionNames, err := os.ReadDir(config.StartDir[0] + "/" + envDir.Name() + "/" + fileSteppedInto.Name() + "/" + projectDirectory.Name())
 						if err != nil {
 							config.Log.Printf("Sections Couldn't read into: %s \n", projectDirectory.Name())
 						}
@@ -306,7 +304,7 @@ func SeedVault(config *eUtils.DriverConfig) error {
 								continue
 							}
 
-							sectionConfigFiles, err := ioutil.ReadDir(config.StartDir[0] + "/" + envDir.Name() + "/" + fileSteppedInto.Name() + "/" + projectDirectory.Name() + "/" + sectionName.Name())
+							sectionConfigFiles, err := os.ReadDir(config.StartDir[0] + "/" + envDir.Name() + "/" + fileSteppedInto.Name() + "/" + projectDirectory.Name() + "/" + sectionName.Name())
 							if err != nil {
 								config.Log.Printf("Section Config Couldn't read into: %s \n", sectionName.Name())
 							}
@@ -316,7 +314,7 @@ func SeedVault(config *eUtils.DriverConfig) error {
 								if strings.HasPrefix(sectionConfigFile.Name(), ".") || (config.SubSectionValue != "" && (sectionConfigFile.Name() != config.SubSectionValue)) {
 									continue
 								}
-								subSectionConfigFiles, err := ioutil.ReadDir(path)
+								subSectionConfigFiles, err := os.ReadDir(path)
 
 								if err != nil {
 									config.Log.Printf("Sub Sections Couldn't read into: %s \n", config.SubSectionName)
@@ -332,7 +330,7 @@ func SeedVault(config *eUtils.DriverConfig) error {
 										}
 
 										if subSectionConfigFile.IsDir() {
-											deepNestedFiles, err := ioutil.ReadDir(subSectionPath)
+											deepNestedFiles, err := os.ReadDir(subSectionPath)
 											if err != nil {
 												config.Log.Printf("Deep Nested Couldn't read into: %s \n", config.SubSectionName)
 												continue
@@ -341,7 +339,7 @@ func SeedVault(config *eUtils.DriverConfig) error {
 											for _, deepNestedFile := range deepNestedFiles {
 												if deepNestedFile.IsDir() {
 													subSectionPath = subSectionPath + "/" + deepNestedFile.Name()
-													deeplyNestedFiles, err := ioutil.ReadDir(subSectionPath)
+													deeplyNestedFiles, err := os.ReadDir(subSectionPath)
 													if err != nil {
 														config.Log.Printf("Sub secting deep nested Couldn't read into: %s \n", config.SubSectionName)
 														continue
@@ -439,7 +437,7 @@ func SeedVault(config *eUtils.DriverConfig) error {
 
 // SeedVaultFromFile takes a file path and seeds the vault with the seeds found in an individual file
 func SeedVaultFromFile(config *eUtils.DriverConfig, filepath string) {
-	rawFile, err := ioutil.ReadFile(filepath)
+	rawFile, err := os.ReadFile(filepath)
 	// Open file
 	eUtils.LogErrorAndSafeExit(config, err, 1)
 	if config.WantCerts && (strings.Contains(filepath, "/Index/") || strings.Contains(filepath, "/PublicIndex/") || strings.Contains(filepath, "/Restricted/")) {
@@ -475,7 +473,7 @@ func seedVaultWithCertsFromEntry(config *eUtils.DriverConfig, mod *helperkv.Modi
 	}
 	certPath = coreopts.GetFolderPrefix(nil) + "_seeds/" + certPath
 	eUtils.LogInfo(config, fmt.Sprintf("Inspecting certificate: "+certPath+"."))
-	cert, err := ioutil.ReadFile(certPath)
+	cert, err := os.ReadFile(certPath)
 	if err != nil {
 		eUtils.LogErrorObject(config, err, false)
 		return
@@ -513,16 +511,12 @@ func seedVaultWithCertsFromEntry(config *eUtils.DriverConfig, mod *helperkv.Modi
 				fallthrough
 			case "dev":
 				certHost = strings.Replace(certHost, "*", "develop", 1)
-				break
 			case "QA":
 				certHost = strings.Replace(certHost, "*", "qa", 1)
-				break
 			case "auto":
 				certHost = strings.Replace(certHost, "*", "auto", 1)
-				break
 			case "performance":
 				certHost = strings.Replace(certHost, "*", "performance", 1)
-				break
 			}
 
 			opts := x509.VerifyOptions{
@@ -614,7 +608,6 @@ func seedVaultWithCertsFromEntry(config *eUtils.DriverConfig, mod *helperkv.Modi
 		eUtils.LogInfo(config, "Missing expected cert at: "+certPath+".  Cert will not be loaded.")
 		return
 	}
-	return
 }
 
 // SeedVaultFromData takes file bytes and seeds the vault with contained data
@@ -647,7 +640,7 @@ func SeedVaultFromData(config *eUtils.DriverConfig, filepath string, fData []byt
 	}
 
 	seed, ok := rawYaml.(map[interface{}]interface{})
-	if ok == false {
+	if !ok {
 		return eUtils.LogAndSafeExit(config, "Invalid yaml file.  Refusing to continue.", 1)
 	}
 
