@@ -11,7 +11,6 @@ import (
 	"encoding/pem"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"math/rand"
 	"os"
 	"regexp"
@@ -81,14 +80,14 @@ func GetSetEnvAddrContext(env string, envContext string, addrPort string) (strin
 
 	//This will use env by default, if blank it will use context. If context is defined, it will replace context.
 	if env == "" {
-		file, err := ioutil.ReadFile(dirname + configDir)
+		file, err := os.ReadFile(dirname + configDir)
 		if err != nil {
 			fmt.Printf("Could not read the context file due to this %s error \n", err)
 			return "", "", "", err
 		}
 		fileContent := string(file)
 		if fileContent == "" {
-			return "", "", "", errors.New("Could not read the context file")
+			return "", "", "", errors.New("could not read the context file")
 		}
 		if !strings.Contains(fileContent, envContextPrefix) && envContext != "" {
 			var output string
@@ -98,7 +97,7 @@ func GetSetEnvAddrContext(env string, envContext string, addrPort string) (strin
 				output = fileContent + envContextPrefix + envContext + "\n"
 			}
 
-			if err = ioutil.WriteFile(dirname+configDir, []byte(output), 0600); err != nil {
+			if err = os.WriteFile(dirname+configDir, []byte(output), 0600); err != nil {
 				return "", "", "", err
 			}
 			fmt.Println("Context flag has been written out.")
@@ -109,12 +108,12 @@ func GetSetEnvAddrContext(env string, envContext string, addrPort string) (strin
 			if len(result) == 1 {
 				addrPort = result[0]
 			} else {
-				return "", "", "", errors.New("Couldn't find port.")
+				return "", "", "", errors.New("couldn't find port")
 			}
 			currentEnvContext := strings.TrimSpace(fileContent[strings.Index(fileContent, envContextPrefix)+len(envContextPrefix):])
 			if envContext != "" {
 				output := strings.Replace(fileContent, envContextPrefix+currentEnvContext, envContextPrefix+envContext, -1)
-				if err = ioutil.WriteFile(dirname+configDir, []byte(output), 0600); err != nil {
+				if err = os.WriteFile(dirname+configDir, []byte(output), 0600); err != nil {
 					return "", "", "", err
 				}
 				fmt.Println("Context flag has been written out.")
@@ -143,7 +142,7 @@ func TrcshAuth(config *eUtils.DriverConfig) (*TrcShConfig, error) {
 			fmt.Println("No homedir for current user")
 			os.Exit(1)
 		}
-		fileBytes, err := ioutil.ReadFile(dir + "/.kube/config")
+		fileBytes, err := os.ReadFile(dir + "/.kube/config")
 		if err != nil {
 			fmt.Println("No local kube config found...")
 			os.Exit(1)
@@ -228,7 +227,7 @@ func PenseQuery(pense string) (string, error) {
 
 	r, penseErr := c.Pense(ctx, &cap.PenseRequest{Pense: penseCode, PenseIndex: pense})
 	if penseErr != nil {
-		return "", errors.Join(errors.New("Pense error"), penseErr)
+		return "", errors.Join(errors.New("pense error"), penseErr)
 	}
 
 	return r.GetPense(), nil
