@@ -4,7 +4,7 @@ import (
 	"errors"
 	"fmt"
 	"html/template"
-	"io/ioutil"
+	"os"
 	"strings"
 	"text/template/parse"
 
@@ -69,7 +69,7 @@ func ToSeed(config *eUtils.DriverConfig, mod *helperkv.Modifier,
 		templateFile, _ := vcutils.ConfigTemplateRaw(config, mod, templatePathExtended, configuredFilePath, true, project, serviceRaw, false, true, config.ExitOnFailure)
 		newTemplate = string(templateFile)
 	} else {
-		templateFile, err := ioutil.ReadFile(templatePath)
+		templateFile, err := os.ReadFile(templatePath)
 		newTemplate = string(templateFile)
 		if err != nil {
 			return nil, nil, nil, 0, eUtils.LogAndSafeExit(config, err.Error(), -1)
@@ -130,7 +130,7 @@ func GetInitialTemplateStructure(config *eUtils.DriverConfig, templatePathSlice 
 	var templateDepth int
 
 	// Remove the file format from the name of the template file
-	if strings.Index(templatePathSlice[len(templatePathSlice)-1], ".") >= 0 {
+	if strings.Contains(templatePathSlice[len(templatePathSlice)-1], ".") {
 		idxFileFormat := strings.Index(templatePathSlice[len(templatePathSlice)-1], ".")
 		templatePathSlice[len(templatePathSlice)-1] = templatePathSlice[len(templatePathSlice)-1][:idxFileFormat]
 	}
@@ -228,7 +228,7 @@ func Parse(config *eUtils.DriverConfig, cds *vcutils.ConfigDataStore,
 		if templatePathSlice[templateDir+1] == "Common" {
 			fileOffsetIndex = 2
 		}
-		keyPath := templatePathSlice[templateDir+fileOffsetIndex : len(templatePathSlice)]
+		keyPath := templatePathSlice[templateDir+fileOffsetIndex:]
 
 		AppendToTemplateSection(interfaceTemplateSection,
 			valueSection,
@@ -271,7 +271,7 @@ func Parse(config *eUtils.DriverConfig, cds *vcutils.ConfigDataStore,
 		if len(templatePathSlice)-1 < keyFileIndex {
 			keyFileIndex = len(templatePathSlice) - 1
 		}
-		keyPath := templatePathSlice[keyFileIndex:len(templatePathSlice)]
+		keyPath := templatePathSlice[keyFileIndex:]
 		secret := defaultSecret
 		if keyName == "certData" {
 			secret = "data"
@@ -306,7 +306,7 @@ func Parse(config *eUtils.DriverConfig, cds *vcutils.ConfigDataStore,
 			}
 		}
 	} else {
-		parseMsg := fmt.Sprintf("Template: %s Incorrect template element count: %d Syntax error: %v", templatePathSlice[templateDir+3:len(templatePathSlice)], len(args), args)
+		parseMsg := fmt.Sprintf("Template: %s Incorrect template element count: %d Syntax error: %v", templatePathSlice[templateDir+3:], len(args), args)
 		return eUtils.LogAndSafeExit(config, parseMsg, 1)
 	}
 	return nil
