@@ -4,7 +4,6 @@ import (
 	"bufio"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"os"
 	"strings"
@@ -35,7 +34,7 @@ func (c *cert) getConfig(logger *log.Logger, file string) (*cert, error) {
 		return nil, err
 	}
 
-	yamlFile, err := ioutil.ReadFile(userHome + "/.tierceron/" + file)
+	yamlFile, err := os.ReadFile(userHome + "/.tierceron/" + file)
 	if err != nil {
 		logger.Printf("yamlFile.Get err #%v ", err)
 	}
@@ -103,7 +102,7 @@ func AutoAuth(config *DriverConfig,
 		}
 		if appRoleIDPtr == nil || len(*appRoleIDPtr) == 0 || secretIDPtr == nil || len(*secretIDPtr) == 0 {
 			if config.IsShellSubProcess {
-				return errors.New("Required azure deploy approle and secret are missing.")
+				return errors.New("required azure deploy approle and secret are missing")
 			}
 			if _, err := os.Stat(userHome + "/.tierceron/" + appRoleConfig); !os.IsNotExist(err) {
 				exists = true
@@ -247,7 +246,7 @@ func AutoAuth(config *DriverConfig,
 			}
 
 			// Create cert file
-			writeErr := ioutil.WriteFile(userHome+"/.tierceron/config.yml", dump, 0600)
+			writeErr := os.WriteFile(userHome+"/.tierceron/config.yml", dump, 0600)
 			if writeErr != nil {
 				LogInfo(config, fmt.Sprintf("Unable to write file: %v\n", writeErr))
 			}
@@ -290,12 +289,13 @@ func AutoAuth(config *DriverConfig,
 		}
 
 		tokenNamePrefix := "config"
-		if appRoleConfig == "configpub.yml" {
+		switch appRoleConfig {
+		case "configpub.yml":
 			tokenNamePrefix = "vault_pub"
-		} else if appRoleConfig == "configdeploy.yml" {
+		case "configdeploy.yml":
 			tokenNamePrefix = "vault_token_deploy"
 			goto skipswitch
-		} else if appRoleConfig == "deployauth" {
+		case "deployauth":
 			tokenNamePrefix = "vault_token_azuredeploy"
 			goto skipswitch
 		}
@@ -344,7 +344,7 @@ func AutoAuth(config *DriverConfig,
 
 	if len(*tokenNamePtr) > 0 {
 		if len(*appRoleIDPtr) == 0 || len(*secretIDPtr) == 0 {
-			return errors.New("Need both public and secret app role to retrieve token from vault")
+			return errors.New("need both public and secret app role to retrieve token from vault")
 		}
 
 		roleToken, err := v.AppRoleLogin(*appRoleIDPtr, *secretIDPtr)
@@ -361,13 +361,14 @@ func AutoAuth(config *DriverConfig,
 		}
 		mod.RawEnv = "bamboo"
 		mod.Env = "bamboo"
-		if appRoleConfig == "configpub.yml" {
+		switch appRoleConfig {
+		case "configpub.yml":
 			mod.RawEnv = "pub"
 			mod.Env = "pub"
-		} else if appRoleConfig == "configdeploy.yml" {
+		case "configdeploy.yml":
 			mod.RawEnv = "deploy"
 			mod.Env = "deploy"
-		} else if appRoleConfig == "deployauth" {
+		case "deployauth":
 			mod.RawEnv = "azuredeploy"
 			mod.Env = "azuredeploy"
 		}
