@@ -35,12 +35,12 @@ func IsUrlIp(address string) (bool, error) {
 	return false, nil
 }
 
-//CreateHTTPClient reads from several .pem files to get the necessary keys and certs to configure the http client and returns the client.
+// CreateHTTPClient reads from several .pem files to get the necessary keys and certs to configure the http client and returns the client.
 func CreateHTTPClient(insecure bool, address string, env string, scan bool) (client *http.Client, err error) {
 	return CreateHTTPClientAllowNonLocal(insecure, address, env, scan, false)
 }
 
-//CreateHTTPClient reads from several .pem files to get the necessary keys and certs to configure the http client and returns the client.
+// CreateHTTPClient reads from several .pem files to get the necessary keys and certs to configure the http client and returns the client.
 func CreateHTTPClientAllowNonLocal(insecure bool, address string, env string, scan bool, allowNonLocal bool) (client *http.Client, err error) {
 	// // create a pool of trusted certs
 	certPath := "../../certs/cert_files/dcidevpublic.pem"
@@ -49,7 +49,7 @@ func CreateHTTPClientAllowNonLocal(insecure bool, address string, env string, sc
 	}
 
 	cert, err := Asset(certPath)
-	//servCertPEM, err := ioutil.ReadFile(certPath)
+	//servCertPEM, err := os.ReadFile(certPath)
 	//servCertPEM := []byte(cert)
 	if err != nil {
 		return nil, err
