@@ -15,8 +15,8 @@ func GetStubbedDataFlowStatistics() ([]string, map[string][]float64) {
 	return nil, nil
 }
 
-func BuildFleet(mod *kv.Modifier, logger *log.Logger) (flowcore.TTDINode, error) {
-	return flowcore.TTDINode{}, nil
+func BuildFleet(mod *kv.Modifier, logger *log.Logger) (*flowcore.TTDINode, error) {
+	return &flowcore.TTDINode{}, nil
 }
 
 func GetDataFlowGroups(mod *kv.Modifier, argosy *flowcore.TTDINode) []flowcore.TTDINode {
