@@ -177,7 +177,7 @@ func (m *Modifier) ValidateEnvironment(environment string, init bool, policySuff
 	env, sub, _, envErr := PreCheckEnvironment(environment)
 
 	if envErr != nil {
-		logger.Println(fmt.Sprintf("Environment format error: %v\n", envErr))
+		logger.Printf("Environment format error: %v\n", envErr)
 		return false, envErr
 	} else {
 		if sub != "" {
@@ -197,7 +197,7 @@ func (m *Modifier) ValidateEnvironment(environment string, init bool, policySuff
 	secret, err := m.client.Auth().Token().LookupSelf()
 
 	if err != nil {
-		logger.Println(fmt.Sprintf("LookupSelf Auth failure: %v\n", err))
+		logger.Printf("LookupSelf Auth failure: %v\n", err)
 		if urlErr, urlErrOk := err.(*url.Error); urlErrOk {
 			if _, sErrOk := urlErr.Err.(*tls.CertificateVerificationError); sErrOk {
 				return false, err
@@ -391,7 +391,7 @@ retryVaultAccess:
 							//memprotectopts.MemProtect(nil, &dataValueString)
 							// don't lock but accept json.Number.
 						} else {
-							return nil, errors.New(fmt.Sprintf("Unexpected datatype. Refusing to read what we cannot lock. Nested. %T", dataValues))
+							return nil, fmt.Errorf("unexpected datatype. Refusing to read what we cannot lock. Nested. %T", dataValues)
 						}
 					}
 				} else if dataValueString, isString := dataValues.(string); isString {
@@ -409,13 +409,13 @@ retryVaultAccess:
 					//memprotectopts.MemProtect(nil, &dataValueString)
 					// don't lock but accept json.Number.
 				} else {
-					return nil, errors.New(fmt.Sprintf("Unexpected datatype. Refusing to read what we cannot lock. %T", dataValues))
+					return nil, fmt.Errorf("unexpected datatype. Refusing to read what we cannot lock. %T", dataValues)
 				}
 			}
 		}
 		return data, err
 	}
-	return nil, errors.New("Could not get data from vault response")
+	return nil, errors.New("could not get data from vault response")
 }
 
 // ReadMapValue takes a valueMap, path, and a key and returns the corresponding value from the vault
@@ -429,10 +429,10 @@ func (m *Modifier) ReadMapValue(valueMap map[string]interface{}, path string, ke
 		} else if stringer, ok := valueMap[key].((json.Number)); ok {
 			return stringer.String(), nil
 		} else {
-			return "", fmt.Errorf("Cannot convert value at %s to string", key)
+			return "", fmt.Errorf("cannot convert value at %s to string", key)
 		}
 	}
-	return "", fmt.Errorf("Key '%s' not found in '%s' with env '%s'", key, path, m.Env)
+	return "", fmt.Errorf("key '%s' not found in '%s' with env '%s'", key, path, m.Env)
 }
 
 // ReadValue takes a path and a key and returns the corresponding value from the vault
@@ -470,13 +470,13 @@ retryQuery:
 		}
 	}
 	if err != nil {
-		logger.Printf("Modifier failing after %d retries.\n", retries)
+		logger.Printf("modifier failing after %d retries.\n", retries)
 	}
 
 	if data, ok := secret.Data["metadata"].(map[string]interface{}); ok {
 		return data, err
 	}
-	return nil, errors.New("Could not get metadata from vault response")
+	return nil, errors.New("could not get metadata from vault response")
 }
 
 // ReadVersionMetadata Reads the Metadata of all versions from the path referenced by this Modifier
@@ -505,7 +505,7 @@ retryQuery:
 			goto retryQuery
 		}
 	} else if secret == nil || secret.Data["versions"] == nil {
-		return nil, errors.New("No version data.")
+		return nil, errors.New("no version data")
 	}
 
 	if err != nil {
@@ -515,7 +515,7 @@ retryQuery:
 	if versionsData, ok := secret.Data["versions"].(map[string]interface{}); ok {
 		return versionsData, err
 	}
-	return nil, errors.New("Could not get metadata of versions from vault response")
+	return nil, errors.New("could not get metadata of versions from vault response")
 }
 
 // List lists the paths underneath this one
@@ -728,7 +728,7 @@ func (m *Modifier) GetVersionValues(mod *Modifier, wantCerts bool, enginePath st
 			if _, ok := versionDataMap[certPath]; !ok {
 				metadataValue, err := mod.ReadVersionMetadata(certPath, logger)
 				if err != nil {
-					err := fmt.Errorf("Unable to fetch data from %s", certPath)
+					err := fmt.Errorf("unable to fetch data from %s", certPath)
 					return nil, err
 				}
 				if len(metadataValue) != 0 {
@@ -756,9 +756,7 @@ func (m *Modifier) GetVersionValues(mod *Modifier, wantCerts bool, enginePath st
 					path = enginePath + "/" + path
 					if mod.SubSectionName != "" {
 						subSectionName := mod.SubSectionName
-						if strings.HasPrefix(subSectionName, "/") {
-							subSectionName = subSectionName[1:]
-						}
+						subSectionName = strings.TrimPrefix(subSectionName, "/")
 						path = path + subSectionName
 					}
 
@@ -779,7 +777,7 @@ func (m *Modifier) GetVersionValues(mod *Modifier, wantCerts bool, enginePath st
 
 	mod.Env = realEnv
 	if len(versionDataMap) < 1 {
-		return nil, fmt.Errorf("No version data available for this env")
+		return nil, fmt.Errorf("no version data available for this env")
 	}
 	return versionDataMap, nil
 }
@@ -808,6 +806,9 @@ func (m *Modifier) recursivePathFinder(filePaths []string, versionDataMap map[st
 		}
 
 		metadataValue, err := m.ReadVersionMetadata(filePath, logger)
+		if err != nil {
+			logger.Println(err.Error())
+		}
 		if len(metadataValue) == 0 {
 			continue
 		}
@@ -821,7 +822,7 @@ func (m *Modifier) getPaths(pathName string, logger *log.Logger) ([]string, erro
 	//logger.Println(secrets)
 	pathList := []string{}
 	if err != nil {
-		return nil, fmt.Errorf("Unable to list paths under %s in %s", pathName, m.Env)
+		return nil, fmt.Errorf("unable to list paths under %s in %s", pathName, m.Env)
 	} else if secrets != nil {
 		//add paths
 		slicey := secrets.Data["keys"].([]interface{})
@@ -845,7 +846,7 @@ func (m *Modifier) GetTemplateFilePaths(pathName string, logger *log.Logger) ([]
 	secrets, err := m.List(pathName, logger)
 	pathList := []string{}
 	if err != nil {
-		return nil, fmt.Errorf("Unable to list paths under %s in %s", pathName, m.Env)
+		return nil, fmt.Errorf("unable to list paths under %s in %s", pathName, m.Env)
 	} else if secrets != nil {
 		//add paths
 		slicey := secrets.Data["keys"].([]interface{})
@@ -859,10 +860,7 @@ func (m *Modifier) GetTemplateFilePaths(pathName string, logger *log.Logger) ([]
 		subPathList := []string{}
 		for _, path := range pathList {
 			subsubList, _ := m.templateFileRecurse(path, logger)
-			for _, subsub := range subsubList {
-				//List is returning both pathEnd and pathEnd/
-				subPathList = append(subPathList, subsub)
-			}
+			subPathList = append(subPathList, subsubList...)
 		}
 		if len(subPathList) != 0 {
 			return subPathList, nil
@@ -881,12 +879,10 @@ func (m *Modifier) templateFileRecurse(pathName string, logger *log.Logger) ([]s
 			for _, pathEnd := range subslice {
 				//List is returning both pathEnd and pathEnd/
 				subpath := pathName + pathEnd.(string)
-				subsublist, _ := m.templateFileRecurse(subpath, logger)
-				if len(subsublist) != 0 {
-					for _, subsub := range subsublist {
-						//List is returning both pathEnd and pathEnd/
-						subPathList = append(subPathList, subsub)
-					}
+				subsubList, _ := m.templateFileRecurse(subpath, logger)
+				if len(subsubList) != 0 {
+					//List is returning both pathEnd and pathEnd/
+					subPathList = append(subPathList, subsubList...)
 				}
 				subPathList = append(subPathList, subpath)
 			}
@@ -990,7 +986,7 @@ retryQuery:
 	if secret == nil && err == nil {
 		return nil, nil
 	}
-	return nil, errors.New("Could not get metadata from vault response")
+	return nil, errors.New("could not get metadata from vault response")
 }
 
 func (m *Modifier) HardDelete(path string, logger *log.Logger) (map[string]interface{}, error) {
@@ -1050,5 +1046,5 @@ retryQuery:
 			return nil, err
 		}
 	}
-	return nil, errors.New("Could not get metadata from vault response")
+	return nil, errors.New("could not get metadata from vault response")
 }
