@@ -99,7 +99,7 @@ func (s *Server) InitVault(ctx context.Context, req *pb.InitReq) (*pb.InitResp, 
 	envStrings := SelectedEnvironment
 	for _, e := range envStrings {
 		mod.Env = e
-		err, warn = il.UploadTemplateDirectory(mod, templatePath, logger)
+		warn, err = il.UploadTemplateDirectory(mod, templatePath, logger)
 		eUtils.LogErrorObject(config, err, false)
 		eUtils.LogWarningsObject(config, warn, false)
 	}
